requirements.txt
PyYAML>=3.10
six
yapsy==1.11.223